+++ src/main/java/me/desht/pneumaticcraft/common/thirdparty/jei/JEIUVLightBoxCategory.java
@@ -19,7 +19,6 @@ package me.desht.pneumaticcraft.common.thirdparty.jei;
 
 import me.desht.pneumaticcraft.common.block.entity.UVLightBoxBlockEntity;
 import me.desht.pneumaticcraft.common.core.ModBlocks;
-import me.desht.pneumaticcraft.common.core.ModItems;
 import me.desht.pneumaticcraft.common.item.EmptyPCBItem;
 import me.desht.pneumaticcraft.common.recipes.machine.UVLightBoxRecipe;
 import me.desht.pneumaticcraft.lib.Textures;
@@ -30,14 +29,13 @@ import mezz.jei.api.gui.drawable.IDrawableStatic;
 import mezz.jei.api.gui.ingredient.IRecipeSlotsView;
 import mezz.jei.api.recipe.IFocusGroup;
 import mezz.jei.api.recipe.RecipeIngredientRole;
+import net.minecraft.Util;
 import net.minecraft.client.gui.GuiGraphics;
 import net.minecraft.network.chat.Component;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.crafting.Ingredient;
 import net.minecraftforge.registries.ForgeRegistries;
 
-import java.util.ArrayList;
-import java.util.Collections;
 import java.util.List;
 
 import static me.desht.pneumaticcraft.common.util.PneumaticCraftUtils.xlate;
@@ -45,21 +43,13 @@ import static me.desht.pneumaticcraft.common.util.PneumaticCraftUtils.xlate;
 public class JEIUVLightBoxCategory extends AbstractPNCCategory<UVLightBoxRecipe> {
     private final IDrawableAnimated progressBar;
 
-    private static final List<UVLightBoxRecipe> UV_LIGHT_BOX_RECIPES;
-    static {
-        List<UVLightBoxRecipe> recipes = new ArrayList<>();
-        ForgeRegistries.ITEMS.getValues().stream()
-                .filter(item -> item instanceof EmptyPCBItem)
-                .forEach(item -> {
-                    ItemStack out = new ItemStack(item);
-                    ItemStack in = new ItemStack(item);
-                    UVLightBoxBlockEntity.setExposureProgress(out, 100);
-                    UVLightBoxRecipe recipe = new UVLightBoxRecipe(Ingredient.of(in), out);
-                    recipes.add(recipe);
-                });
-        // add recipes to list
-        UV_LIGHT_BOX_RECIPES = Collections.unmodifiableList(recipes);
-    }
+    private static final List<UVLightBoxRecipe> UV_LIGHT_BOX_RECIPES = ForgeRegistries.ITEMS.getValues().stream()
+            .filter(item -> item instanceof EmptyPCBItem)
+            .map(item -> new UVLightBoxRecipe(
+                    Ingredient.of(item),
+                    Util.make(new ItemStack(item), s -> UVLightBoxBlockEntity.setExposureProgress(s, 100))
+            ))
+            .toList();
 
     JEIUVLightBoxCategory() {
         super(RecipeTypes.UV_LIGHT_BOX,